requirements.txt
flask
pymongo
flask-cors
python-dotenv
flask-session
bcrypt
waitress
gunicorn
